U main.go, A static_*.go: fix no-embed build: split static and support build tag embed
